little refactor (#125)
* check verify time based 2fa

* included the verify route
