+++ src/controllers/auth/index.ts
@@ -1,13 +1,12 @@
-export * from './completeTimeBased2faSetup';
 export * from './forgotPassword';
 export * from './protect';
 export * from './resendVerification';
 export * from './resetPassword';
 export * from './session';
-export * from './setupTimeBased2fa';
+export * from './setup2fa';
 export * from './signin';
 export * from './signout';
 export * from './signup';
 export * from './verifyEmail';
-export * from './get2faCodeViaEmail';
+export * from './setup2faCodeViaEmail';
 export * from './verifyTimeBased2fa';
