--- src/controllers/auth/setupTimeBased2fa.ts
@@ -1,85 +1,0 @@
-import { catchAsync } from '@/middlewares';
-import { Request, Response } from 'express';
-import AppError from '@/common/utils/appError';
-import {
-	AppResponse,
-	generateRandom6DigitKey,
-	generateRandomBase32,
-	generateTimeBased2fa,
-	getFromCache,
-	hashData,
-	setCache,
-	toJSON,
-} from '@/common/utils';
-import { UserModel } from '@/models';
-import { Require_id } from 'mongoose';
-import { IUser } from '../../common/interfaces';
-import { twoFactorTypeEnum } from '../../common/constants';
-import { addEmailToQueue } from '../../queues/emailQueue';
-
-export const setupTimeBased2fa = catchAsync(async (req: Request, res: Response) => {
-	const { user } = req;
-	const { twoFactorType } = req.body;
-
-	if (!twoFactorType) {
-		throw new AppError('Invalid Request', 400);
-	}
-
-	if (!user) {
-		throw new AppError('Unauthorized', 401);
-	}
-
-	if (user?.timeBased2FA && user.timeBased2FA.active) {
-		throw new AppError('2FA is already active', 400);
-	}
-
-	if (twoFactorType === twoFactorTypeEnum.EMAIL) {
-		const token = generateRandom6DigitKey();
-		const hashedToken = hashData({ token }, { expiresIn: '5m' });
-
-		await addEmailToQueue({
-			type: 'get2faCodeViaEmail',
-			data: {
-				to: user.email,
-				name: user.firstName,
-				twoFactorCode: token,
-				expiryTime: '5',
-				priority: 'high',
-			},
-		});
-
-		await setCache(`2FAEmailCode:${user._id.toString()}`, { token: hashedToken }, 300);
-
-		return AppResponse(res, 200, null, 'OTP code sent to email successfully');
-	}
-
-	const secret = generateRandomBase32();
-	const qrCode = await generateTimeBased2fa(secret);
-	const hashedSecret = hashData({ token: secret }, { expiresIn: 0 });
-
-	await UserModel.findByIdAndUpdate(user?._id, {
-		timeBased2FA: {
-			secret: hashedSecret,
-		},
-	});
-
-	const userFromCache = await getFromCache<Require_id<IUser>>(user._id.toString());
-
-	if (userFromCache) {
-		// update cache
-		await setCache(
-			user._id.toString()!,
-			toJSON({ ...userFromCache, timeBased2FA: { secret: hashedSecret, active: false } }, [])
-		);
-	}
-
-	return AppResponse(
-		res,
-		200,
-		{
-			secret,
-			qrCode,
-		},
-		'Created 2FA successfully'
-	);
-});
