+++ src/models/userModel.ts
@@ -163,7 +163,6 @@ userSchema.pre('save', async function (next) {
 			this.email,
 			this.phoneNumber,
 			this.photo,
-			//this.address.length,
 			this.gender,
 			this.isIdVerified,
 			this.isMobileVerified,
