+++ src/routes/authRouter.ts
@@ -1,5 +1,4 @@
 import {
-	completeTimeBased2fa,
 	get2faCodeViaEmail,
 	forgotPassword,
 	protect,
@@ -13,6 +12,7 @@ import {
 	verifyEmail,
 	verifyTimeBased2fa,
 } from '@/controllers';
+import { completeTimeBased2fa } from '@/controllers/auth/completeTimeBased2faSetup';
 import { Router } from 'express';
 
 const router = Router();
@@ -31,5 +31,6 @@ router.get('/session', session);
 router.get('/signout', signOut);
 router.post('/2fa/time/setup', setupTimeBased2fa);
 router.post('/2fa/time/complete', completeTimeBased2fa);
+router.post('/2fa/time/verify', verifyTimeBased2fa);
 
 export { router as authRouter };
